+++ src/api/table/list.ts
@@ -1,4 +1,4 @@
-import http from '@/utils/http/axios';
+import { http } from '@/utils/http/axios';
 
 //获取table
 export function getTableList(params) {